+++ vertx-core/src/main/java/io/vertx/core/http/impl/Http1xClientConnection.java
@@ -37,7 +37,6 @@ import io.vertx.core.http.HttpMethod;
 import io.vertx.core.http.HttpVersion;
 import io.vertx.core.http.WebSocketVersion;
 import io.vertx.core.http.impl.headers.HeadersAdaptor;
-import io.vertx.core.http.impl.http2.Http2ClientPush;
 import io.vertx.core.internal.ContextInternal;
 import io.vertx.core.internal.PromiseInternal;
 import io.vertx.core.internal.buffer.BufferInternal;
@@ -524,7 +523,7 @@ public class Http1xClientConnection extends Http1xConnection implements HttpClie
     }
 
     @Override
-    public HttpClientStream pushHandler(Handler<Http2ClientPush> handler) {
+    public HttpClientStream pushHandler(Handler<HttpClientPush> handler) {
       // No op
       return this;
     }
