+++ vertx-core/src/main/java/io/vertx/core/http/impl/StatisticsGatheringHttpClientStream.java
@@ -19,7 +19,6 @@ import io.vertx.core.buffer.Buffer;
 import io.vertx.core.http.HttpFrame;
 import io.vertx.core.http.HttpVersion;
 import io.vertx.core.http.StreamPriority;
-import io.vertx.core.http.impl.http2.Http2ClientPush;
 import io.vertx.core.internal.ContextInternal;
 import io.vertx.core.net.endpoint.ServerInteraction;
 
@@ -103,7 +102,7 @@ class StatisticsGatheringHttpClientStream implements HttpClientStream {
   }
 
   @Override
-  public HttpClientStream pushHandler(Handler<Http2ClientPush> handler) {
+  public HttpClientStream pushHandler(Handler<HttpClientPush> handler) {
     delegate.pushHandler(handler);
     return this;
   }
